+++ meilisearch/src/routes/indexes/documents.rs
@@ -279,7 +279,7 @@ async fn document_addition(
     .await;
 
     let documents_count = match documents_count {
-        Ok(Ok(documents_count)) => documents_count as u64,
+        Ok(Ok(documents_count)) => documents_count,
         // in this case the file has not possibly be persisted.
         Ok(Err(e)) => return Err(e),
         Err(e) => {
